updated amenities and beaches tables

+++ db/migrate/20150629201312_create_beaches.rb
@@ -7,7 +7,7 @@ class CreateBeaches < ActiveRecord::Migration
       t.string :state, null: false
       t.string :zip, null: false
       t.string :name, null: false
-      t.integer :entrance_fee
+      t.decimal :entrance_fee
       t.string :picture_url
       t.text :description
       t.integer :user_id, null: false
